Add production settings

+++ backend/index.js
@@ -1,8 +1,12 @@
 const app = require("./app");
 const { sequelize } = require("./db");
-const { PORT } = require("./utils/env");
+const { PORT, NODE_ENV } = require("./utils/env");
 
 app.listen(PORT, async () => {
-  await sequelize.sync({ force: true }); // this will drop every table
+  if (NODE_ENV === "PRODUCTION") {
+    await sequelize.sync();
+  } else {
+    await sequelize.sync({ force: true }); // this will drop every table
+  }
   console.log(`listening on http://localhost:${PORT}`);
 });
